refactor: module structure for peermap
