+++ iroh-net/src/magicsock/peer_map/best_addr.rs
